requirements.txt
djangorestframework==3.12.4
python-dateutil==2.8.0